user search and admin api made
