+++ api/src/repositories/index.ts
@@ -1,3 +1,4 @@
+export * from './admin/admin.repository';
 export * from './avatar/avatar.repository';
 export * from './block/block.repository';
 export * from './colony/colony.repository';
