+++ api/src/routes/admin.routes.ts
@@ -1,0 +1,8 @@
+import Router from 'express';
+import {adminController} from '../controllers';
+
+const adminRoutes = Router();
+adminRoutes.post('/search', (request, response) =>
+  adminController.searchUsers(request, response));
+
+export {adminRoutes};
